refactor: extract setupParams() to utils.js
- This is done to make setupParams() modular
